Fix generateNumber function parameter types to accept double values for better precision

+++ src/utils.cpp
@@ -8,7 +8,7 @@ void utils::clear_screen() {
 #endif
 }
 
-double utils::generateNumber(int lowerBound, int higherBound) {
+double utils::generateNumber(double lowerBound, double higherBound) {
   std::default_random_engine randomEngine(
       std::chrono::system_clock::now().time_since_epoch().count());
   std::uniform_real_distribution<double> generateNumber(lowerBound,
